+++ hat/assets/js/apps/Iaso/domains/app/translations/en.json
@@ -15,7 +15,6 @@
     "iaso.algo.addRun": "Add algorithm run",
     "iaso.archived.title": "Archives",
     "iaso.area": "Health area",
-    "iaso.restricted_submissions_by_orgunits": "You're only allowed to see submissions for the following orgunits:",
     "iaso.completeness.generateDerivedInstances": "Generate derived instances",
     "iaso.completeness.title": "Completeness",
     "iaso.dataSources.continueOnError": "Continue on error",
@@ -79,7 +78,6 @@
     "iaso.devices.timeSynched": "Time synchronised",
     "iaso.devices.timeUpdated": "Date modified",
     "iaso.entities.create": "Create entity",
-    "iaso.entities.dialog.deleteText": "This operation cannot be undone.",
     "iaso.entities.dialog.deleteTitle": "Are you sure you want to delete this entity?",
     "iaso.entities.title": "Entities",
     "iaso.entities.type": "Type",
@@ -87,7 +85,6 @@
     "iaso.entities.update": "Update entity",
     "iaso.entityTypes.count": "Entities count",
     "iaso.entityTypes.create": "Create entity type",
-    "iaso.entityTypes.dialog.deleteText": "This operation cannot be undone.",
     "iaso.entityTypes.dialog.deleteTitle": "Are you sure you want to delete this entity type?",
     "iaso.entityTypes.title": "Entity types",
     "iaso.entityTypes.update": "Update entity type",
@@ -227,6 +224,7 @@
     "iaso.label.dateFrom": "Start date",
     "iaso.label.dateTo": "End date",
     "iaso.label.delete": "Delete",
+    "iaso.label.deleteText": "This operation cannot be undone.",
     "iaso.label.deleted": "Deleted",
     "iaso.label.derived": "Deduced from another form",
     "iaso.label.destination": "Destination",
@@ -538,6 +536,7 @@
     "iaso.projects.true": "User needs authentication",
     "iaso.projects.update": "Update project",
     "iaso.province": "Province",
+    "iaso.restricted_submissions_by_orgunits": "You're only allowed to see submissions for the following orgunits:",
     "iaso.runs.dialog.deleteRunText": "All created links will be deleted.",
     "iaso.runs.dialog.deleteRunTitle": "Are you sure you want to delete this run?",
     "iaso.search": "Search",
