+++ hat/assets/js/apps/Iaso/domains/app/translations/fr.json
@@ -15,7 +15,6 @@
     "iaso.algo.addRun": "Exécuter un algorithme",
     "iaso.archived.title": "Archives",
     "iaso.area": "Aire de santé",
-    "iaso.restricted_submissions_by_orgunits": "Vous êtes autorisé(e) à voir uniquement les unités d'organisation suivantes:",
     "iaso.completeness.generateDerivedInstances": "Générer les instances dérivées",
     "iaso.completeness.title": "Complétude des données",
     "iaso.dataSources.continueOnError": "Continuer en cas d'erreur",
@@ -76,7 +75,6 @@
     "iaso.devices.timeSynched": "Temps de synchronisation",
     "iaso.devices.timeUpdated": "Dernière modification",
     "iaso.entities.create": "Créer une entité",
-    "iaso.entities.dialog.deleteText": "Cette opération est définitive.",
     "iaso.entities.dialog.deleteTitle": "Êtes-vous certain de vouloir effacer cette entité ?",
     "iaso.entities.title": "Entités",
     "iaso.entities.type": "Type",
@@ -84,7 +82,6 @@
     "iaso.entities.update": "Mettre l'entité à jour",
     "iaso.entityTypes.count": "Nombre d'entités",
     "iaso.entityTypes.create": "Créer un type d'entité",
-    "iaso.entityTypes.dialog.deleteText": "Cette opération est définitive.",
     "iaso.entityTypes.dialog.deleteTitle": "Êtes-vous certain de vouloir effacer ce type d'entité ?",
     "iaso.entityTypes.title": "Types d'entité",
     "iaso.entityTypes.update": "Mettre le type d'entité à jour",
@@ -224,6 +221,7 @@
     "iaso.label.dateFrom": "Date de début",
     "iaso.label.dateTo": "Date de fin",
     "iaso.label.delete": "Effacer",
+    "iaso.label.deleteText": "Cette opération est définitive.",
     "iaso.label.deleted": "Suppression",
     "iaso.label.derived": "Instances dérivées d'un autre formulaire (Pas de collecte mobile possible)",
     "iaso.label.destination": "Destination",
@@ -243,7 +241,6 @@
     "iaso.label.ended_at": "Fin",
     "iaso.label.execute": "Exécuter",
     "iaso.label.export": "Export",
-    "iaso.orgUnits.bulkChangeCount": "Vous allez modifier {count} unités d'org.",
     "iaso.label.exportRequests": "Demandes d'export",
     "iaso.label.exportSelection": "Exporter {count} soumissions",
     "iaso.label.featureFlags": "Options",
@@ -445,6 +442,7 @@
     "iaso.orgUnits.addOrgUnitType": "Ajouter un type d'unité",
     "iaso.orgUnits.addSource": "Ajouter une source",
     "iaso.orgUnits.addToGroups": "Ajouter au(x) groupe(s)",
+    "iaso.orgUnits.bulkChangeCount": "Vous allez modifier {count} unités d'org.",
     "iaso.orgUnits.confirmMultiChange": "Confirmer les modifications multiples ?",
     "iaso.orgUnits.dataSources": "Sources de données",
     "iaso.orgUnits.depth": "Niveau",
@@ -527,6 +525,7 @@
     "iaso.projects.true": "L'utilisateur a besoin d'une authentification",
     "iaso.projects.update": "Modifier un projet",
     "iaso.province": "Province",
+    "iaso.restricted_submissions_by_orgunits": "Vous êtes autorisé(e) à voir uniquement les unités d'organisation suivantes:",
     "iaso.runs.dialog.deleteRunText": "Toutes les correspondances créées seront effacées",
     "iaso.runs.dialog.deleteRunTitle": "Etes-vous certain de vouloir effacer cet exécution d'algorithme ?",
     "iaso.search": "Rechercher",
