+++ hat/assets/js/apps/Iaso/domains/entities/components/Filters.tsx
@@ -1,4 +1,4 @@
-import React, { useState, FunctionComponent } from 'react';
+import React, { useState, FunctionComponent, useCallback } from 'react';
 import { useDispatch } from 'react-redux';
 
 import { Grid, Button, makeStyles } from '@material-ui/core';
@@ -12,7 +12,7 @@ import MESSAGES from '../messages';
 
 import { baseUrl } from '../config';
 
-import { useGetTypes } from '../entityTypes/hooks/useGetTypes';
+import { useGetTypes } from '../entityTypes/hooks/requests';
 
 const useStyles = makeStyles(theme => ({
     ...commonStyles(theme),
@@ -32,7 +32,7 @@ type Props = {
 
 const Filters: FunctionComponent<Props> = ({ params }) => {
     const [filtersUpdated, setFiltersUpdated] = useState(false);
-    const classes: any = useStyles();
+    const classes: Record<string, string> = useStyles();
     const { formatMessage } = useSafeIntl();
     const dispatch = useDispatch();
     const [filters, setFilters] = useState({
@@ -42,7 +42,8 @@ const Filters: FunctionComponent<Props> = ({ params }) => {
 
     const { data: entityTypes, isFetching: fetchingEntitytypes } =
         useGetTypes();
-    const handleSearch = () => {
+
+    const handleSearch = useCallback(() => {
         if (filtersUpdated) {
             setFiltersUpdated(false);
             const tempParams = {
@@ -52,14 +53,19 @@ const Filters: FunctionComponent<Props> = ({ params }) => {
             tempParams.page = '1';
             dispatch(redirectTo(baseUrl, tempParams));
         }
-    };
-    const handleChange = (key, value) => {
-        setFiltersUpdated(true);
-        setFilters({
-            ...filters,
-            [key]: value,
-        });
-    };
+    }, [filtersUpdated, dispatch, filters, params]);
+
+    const handleChange = useCallback(
+        (key, value) => {
+            setFiltersUpdated(true);
+            setFilters({
+                ...filters,
+                [key]: value,
+            });
+        },
+        [filters],
+    );
+
     return (
         <>
             <Grid container spacing={4}>
@@ -122,4 +128,4 @@ const Filters: FunctionComponent<Props> = ({ params }) => {
     );
 };
 
-export default Filters;
+export { Filters };
