--- hat/assets/js/apps/Iaso/domains/entities/hooks/useDelete.ts
@@ -1,12 +1,0 @@
-import { UseMutationResult } from 'react-query';
-import { useSnackMutation } from '../../../libs/apiHooks';
-import { deleteRequest } from '../../../libs/Api';
-import MESSAGES from '../messages';
-
-export const useDelete = (): UseMutationResult =>
-    useSnackMutation(
-        body => deleteRequest(`/api/entity/${body.id}/`),
-        MESSAGES.deleteSuccess,
-        MESSAGES.deleteError,
-        ['entities'],
-    );
