+++ hat/assets/js/apps/Iaso/types/intl.ts
@@ -1,0 +1,5 @@
+export type IntlMessage = {
+    id: string;
+    defaultMessage: string;
+    values?: Record<string, any>;
+};
