+++ hat/assets/js/apps/Iaso/types/table.ts
@@ -1,0 +1,19 @@
+import { ReactElement } from 'react';
+
+export type Column = {
+    Header: string;
+    id?: string;
+    accessor: string;
+    sortable?: boolean;
+    resizable?: boolean;
+    // eslint-disable-next-line no-unused-vars
+    Cell?: (s: any) => ReactElement;
+};
+export type Pagination = {
+    pages: number;
+    page: number;
+    count: number;
+    limit: number;
+    has_next: boolean;
+    has_previous: boolean;
+};
